+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-##
+## 1.10.0
+- Adopt esp-serial-flasher version number
+- Add `erase_flash` command ([#6](https://github.com/higaski/QtESPFlasher/pull/6))
 - Update to esp-serial-flasher 1.10.0
 - Update to esptool-legacy-flasher-stub 1.9.0